feat: move model to component
